Fixed issue #22

+++ src/ru/gelin/android/sendtosd/progress/SizeUnit.java
@@ -4,15 +4,15 @@ import ru.gelin.android.sendtosd.R;
 
 public enum SizeUnit {
     
-    MEGABYTE(1024 * 1024, 1 * 1024 * 1024, R.string.mb_progress),
-    KILOBYTE(1024, 1 * 1024, R.string.kb_progress),
-    BYTE(1, 1, R.string.b_progress),
-    NULL(1, 0, R.string.dummy);
+    MEGABYTE(1024f * 1024, 1 * 1024 * 1024, R.string.mb_progress),
+    KILOBYTE(1024f, 1 * 1024, R.string.kb_progress),
+    BYTE(1f, 1, R.string.b_progress),
+    NULL(1f, 0, R.string.dummy);
     
     /**
      *  How many bytes in one unit?
      */
-    final long multiplier;
+    final float multiplier;
     /**
      *  Limit of the file size, starting from which to use this unit.
      */
@@ -22,7 +22,7 @@ public enum SizeUnit {
      */
     final int progressString;
     
-    SizeUnit(long multiplier, long limit, int progressString) {
+    SizeUnit(float multiplier, long limit, int progressString) {
         this.multiplier = multiplier;
         this.limit = limit;
         this.progressString = progressString;
